+++ fabric-convention-tags-v2/src/main/resources/fabric-convention-tags-api-v2.mixins.json
@@ -3,7 +3,7 @@
   "package": "net.fabricmc.fabric.mixin.convention",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "TagKeyExtensionMixin"
+    "TagKeyMixin"
   ],
   "injectors": {
     "defaultRequire": 1
