+++ fabric-convention-tags-v2/src/main/resources/fabric.mod.json
@@ -32,7 +32,7 @@
   "custom": {
     "fabric-api:module-lifecycle": "stable",
     "loom:injected_interfaces": {
-      "net/minecraft/class_6862": ["net/fabricmc/fabric/api/tag/convention/v2/FabricTagKey"]
+      "net/minecraft/class_6862": ["net/fabricmc/fabric/api/tag/FabricTagKey"]
     }
   }
 }
